MCR-2436 create an additional MCRAbstractCondition class
and move fact-based condition into a sub package *.fact
