+++ mycore-base/src/main/resources/config/mycore.properties
@@ -183,16 +183,16 @@ MCR.Access.XML.category=org.mycore.access.xml.conditions.MCRCategoryCondition
 MCR.Access.Facts.Condition.and=org.mycore.access.facts.condition.combined.MCRAndCondition
 MCR.Access.Facts.Condition.or=org.mycore.access.facts.condition.combined.MCROrCondition
 MCR.Access.Facts.Condition.not=org.mycore.access.facts.condition.combined.MCRNotCondition
-MCR.Access.Facts.Condition.id=org.mycore.access.facts.condition.MCRSimpleCondition
-MCR.Access.Facts.Condition.target=org.mycore.access.facts.condition.MCRSimpleCondition
-MCR.Access.Facts.Condition.action=org.mycore.access.facts.condition.MCRSimpleCondition
-MCR.Access.Facts.Condition.user=org.mycore.access.facts.condition.MCRUserCondition
-MCR.Access.Facts.Condition.role=org.mycore.access.facts.condition.MCRRoleCondition
-MCR.Access.Facts.Condition.ip=org.mycore.access.facts.condition.MCRIPCondition
-MCR.Access.Facts.Condition.status=org.mycore.access.facts.condition.MCRStateCondition
-MCR.Access.Facts.Condition.createdBy=org.mycore.access.facts.condition.MCRCreatedByCondition
-MCR.Access.Facts.Condition.regex=org.mycore.access.facts.condition.MCRRegExCondition
-MCR.Access.Facts.Condition.category=org.mycore.access.facts.condition.MCRCategoryCondition
+MCR.Access.Facts.Condition.id=org.mycore.access.facts.condition.fact.MCRSimpleCondition
+MCR.Access.Facts.Condition.target=org.mycore.access.facts.condition.fact.MCRSimpleCondition
+MCR.Access.Facts.Condition.action=org.mycore.access.facts.condition.fact.MCRSimpleCondition
+MCR.Access.Facts.Condition.user=org.mycore.access.facts.condition.fact.MCRUserCondition
+MCR.Access.Facts.Condition.role=org.mycore.access.facts.condition.fact.MCRRoleCondition
+MCR.Access.Facts.Condition.ip=org.mycore.access.facts.condition.fact.MCRIPCondition
+MCR.Access.Facts.Condition.status=org.mycore.access.facts.condition.fact.MCRStateCondition
+MCR.Access.Facts.Condition.createdBy=org.mycore.access.facts.condition.fact.MCRCreatedByCondition
+MCR.Access.Facts.Condition.regex=org.mycore.access.facts.condition.fact.MCRRegExCondition
+MCR.Access.Facts.Condition.category=org.mycore.access.facts.condition.fact.MCRCategoryCondition
 
 ##############################################################################
 # Persistence definitions
